// lib/types.ts
export interface ScrapedProduct {
  name?: string;
  brand?: string;
  gender?: string;
  notes?: string;
  description?: string;
  imageUrl?: string;
  price?: string;
}